backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pypdf
sentence-transformers
faiss-cpu
httpx
